DACT-465 reorder skeleton pages
