+++ src/controllers/director.appointed.date.controller.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { DIRECTOR_OCCUPATION_PATH } from "../types/page.urls";
+import { DIRECTOR_NATIONALITY_PATH, DIRECTOR_OCCUPATION_PATH } from "../types/page.urls";
 import { Templates } from "../types/template.paths";
 import { urlUtils } from "../utils/url";
 
@@ -15,5 +15,6 @@ export const get = async (req: Request, res: Response, next: NextFunction) => {
 };
 
 export const post = async (req: Request, res: Response, next: NextFunction) => {
-  // not yet implemented
+  const nextPageUrl = urlUtils.getUrlToPath(DIRECTOR_NATIONALITY_PATH, req);
+  return res.redirect(nextPageUrl);
 };
