+++ src/controllers/director.occupation.controller.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { DIRECTOR_APPOINTED_DATE_PATH, DIRECTOR_NATIONALITY_PATH } from "../types/page.urls";
+import { DIRECTOR_NATIONALITY_PATH } from "../types/page.urls";
 import { Templates } from "../types/template.paths";
 import { urlUtils } from "../utils/url";
 
@@ -15,6 +15,5 @@ export const get = async (req: Request, res: Response, next: NextFunction) => {
 };
 
 export const post = async (req: Request, res: Response, next: NextFunction) => {
-  const nextPageUrl = urlUtils.getUrlToPath(DIRECTOR_APPOINTED_DATE_PATH, req);
-  return res.redirect(nextPageUrl);
+  // Not yet implemented
 };
