+++ test/controllers/director.appointed.date.controller.unit.ts
@@ -4,7 +4,7 @@ import mocks from "../mocks/all.middleware.mock";
 import request from "supertest";
 import app from "../../src/app";
 
-import { DIRECTOR_APPOINTED_DATE_PATH, urlParams } from "../../src/types/page.urls";
+import { DIRECTOR_APPOINTED_DATE_PATH, DIRECTOR_NATIONALITY_PATH, urlParams } from "../../src/types/page.urls";
 import { isActiveFeature } from "../../src/utils/feature.flag";
 
 const mockIsActiveFeature = isActiveFeature as jest.Mock;
@@ -16,6 +16,10 @@ const SUBMISSION_ID = "55555555";
 const PAGE_HEADING = "When was the director appointed?";
 const ERROR_PAGE_HEADING = "Sorry, there is a problem with this service";
 const DIRECTOR_APPOINTED_DATE_URL = DIRECTOR_APPOINTED_DATE_PATH
+  .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
+  .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
+  .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
+  const DIRECTOR_NATIONALITY_URL = DIRECTOR_NATIONALITY_PATH
   .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
   .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
   .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
@@ -41,6 +45,16 @@ describe("Director name controller tests", () => {
         expect(response.text).toContain(ERROR_PAGE_HEADING);
       });
 
+    });
+
+    describe("post tests", () => {
+  
+      it("Should redirect to director nationality page", async () => {
+        const response = await request(app).post(DIRECTOR_APPOINTED_DATE_URL);
+
+        expect(response.text).toContain("Found. Redirecting to " + DIRECTOR_NATIONALITY_URL);
+      });
+      
     });
 
 });
