+++ test/controllers/director.date.of.birth.controller.unit.ts
@@ -4,7 +4,7 @@ import mocks from "../mocks/all.middleware.mock";
 import request from "supertest";
 import app from "../../src/app";
 
-import { DIRECTOR_DATE_OF_BIRTH_PATH, DIRECTOR_NATIONALITY_PATH, urlParams } from "../../src/types/page.urls";
+import { DIRECTOR_APPOINTED_DATE_PATH, DIRECTOR_DATE_OF_BIRTH_PATH, urlParams } from "../../src/types/page.urls";
 import { isActiveFeature } from "../../src/utils/feature.flag";
 
 const mockIsActiveFeature = isActiveFeature as jest.Mock;
@@ -19,7 +19,7 @@ const DIRECTOR_DATE_OF_BIRTH_URL = DIRECTOR_DATE_OF_BIRTH_PATH
   .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
   .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
   .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
-const DIRECTOR_NATIONALITY_URL = DIRECTOR_NATIONALITY_PATH
+const DIRECTOR_APPOINTED_DATE_URL = DIRECTOR_APPOINTED_DATE_PATH
   .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
   .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
   .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
@@ -50,10 +50,10 @@ describe("Director name controller tests", () => {
 
     describe("post tests", () => {
   
-      it("Should redirect to director nationality page", async () => {
+      it("Should redirect to director appointed date page", async () => {
         const response = await request(app).post(DIRECTOR_DATE_OF_BIRTH_URL);
 
-        expect(response.text).toContain("Found. Redirecting to " + DIRECTOR_NATIONALITY_URL);
+        expect(response.text).toContain("Found. Redirecting to " + DIRECTOR_APPOINTED_DATE_URL);
       });
       
     });
