+++ test/controllers/director.occupation.controller.unit.ts
@@ -4,7 +4,7 @@ import mocks from "../mocks/all.middleware.mock";
 import request from "supertest";
 import app from "../../src/app";
 
-import { DIRECTOR_APPOINTED_DATE_PATH, DIRECTOR_NATIONALITY_PATH, DIRECTOR_OCCUPATION_PATH, urlParams } from "../../src/types/page.urls";
+import { DIRECTOR_OCCUPATION_PATH, urlParams } from "../../src/types/page.urls";
 import { isActiveFeature } from "../../src/utils/feature.flag";
 
 const mockIsActiveFeature = isActiveFeature as jest.Mock;
@@ -16,10 +16,6 @@ const SUBMISSION_ID = "55555555";
 const PAGE_HEADING = "What is the director's occupation?";
 const ERROR_PAGE_HEADING = "Sorry, there is a problem with this service";
 const DIRECTOR_OCCUPATION_URL = DIRECTOR_OCCUPATION_PATH
-  .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
-  .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
-  .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
-const DIRECTOR_APPOINTED_DATE_URL = DIRECTOR_APPOINTED_DATE_PATH
   .replace(`:${urlParams.PARAM_COMPANY_NUMBER}`, COMPANY_NUMBER)
   .replace(`:${urlParams.PARAM_TRANSACTION_ID}`, TRANSACTION_ID)
   .replace(`:${urlParams.PARAM_SUBMISSION_ID}`, SUBMISSION_ID);
@@ -47,13 +43,4 @@ describe("Director name controller tests", () => {
 
     });
 
-    describe("post tests", () => {
-  
-      it("Should redirect to director appointed date page", async () => {
-        const response = await request(app).post(DIRECTOR_OCCUPATION_URL);
-
-        expect(response.text).toContain("Found. Redirecting to " + DIRECTOR_APPOINTED_DATE_URL);
-      });
-      
-    });
 });
